[network] limit number of connection attempts to disconnected peer
Earlier, we would retry to connect to a disconnected peer every 5
seconds, leading to wasted resources and log spew. The new
implementation sets a limit on the number of times the connectivity
manager will tolerate connection failures before giving up on dialing
the peer. It resets the failure counter on successful connection