[types] fix make_placeholder_script.sh
The output format of the compiler has changed and this broken this script & the fixture script. This diff is a temporary fix that makes the script generate the mvbin file again, by extracting strings from the new output, and apply magic to it.

It should be replaced by a proper solution in the future.