[commmon] Restructure canonical serialization
Given the importance of canonical serialization within the code base, we
are restructuring it to make it to simplify access and understanding.

This diff moves each major component into its own separate module (file)
and re-exports so as to not break any dependencies.

In addition, major types are now fully exposed and implemented, namely
signed integers.