Reorganization of Value in the VM.
This diff defines an API for Values that mimics the bytecodes operating on Values.
The idea is to create an abstraction reasonably encapsulated and small to allow for different representations to be tried and tested.

value.rs is the file to look at, as a stand alone file (the diff makes no sense).
Comments in that file should be enough to get a sense. The rest of the changes should show what the clean up is.
You should all be aware of this change given the presentation we had.

There is more work to do, but in order to avoid rebase hell (already done enough) and to step back a bit, I think this is a very good starting point.
It also unblocks people working on Value related diffs (e.g. vector).
I moved canonical (de)serialization into value.rs because I wanted to keep the public API at a minimum.
That could have been accomplished with pub(crate) but I want to review that crate overall so apology for the "intrusive" change.
It could also be argued that the (de)ser code should be in value.rs.

Added a more comprehensive test for values and run cargo test.

Closes: #810
Approved by: tzakian