[terraform] Try to workaround instance type change issue
When the instance type is changed the instance boots with the same EBS
root value on a clean machine. This means that the instance store is
empty, but still referenced in fstab, causing a boot error. This adds
`nofail` to the fstab entry to allow boot to continue.

Closes: #948
Approved by: andll