[language] Split out interpreter entrypoint from function call impl
Previously, the entrypoint into the interpreter loop was through
execute_function_impl. However, the entry into the interpreter is
special -- we need to charge an intrinsic amount of gas only once.

This separates these two out so that we only charge this once instead of
multiple times since this wound up being called interally from the VM as
well.

Closes: #1032
Approved by: runtian-zhou