[Consensus] Verify SyncInfo on demand
Summary:
We used to verify SyncInfo for every incoming messages.
That would incur O(n^2) signature verifications for the vote aggregators in every round, because every vote would carry a SyncInfo with the QCs.
The performance cost of such verification was high and had a visible impact on throughput and latency.

In this diff we're postponing the verification of SyncInfo to a later stage in `EventProcessor::sync_up()` function.
One "unintended" change in this commit is the more explicit way of resetting consensus state during the epoch change.
We used to call the `process_certificate()` function indiscriminatory, and when I changed it to be called only for QCs that are higher than the local,
the epoch transition stopped working: it needed the `SafetyRules::update()` call. Moved the epoch change code of SafetyRules to a separate `start_new_epoch()` function,
but in general I find it kinda fragile. May be we should make consensus state epoch aware (keep a pair of (epoch, round) instead of just round)? That way we wouldn't depend on these reset functions.

Testing: unit tests

Closes: #1759
Approved by: dmitri-perelman