[config] Add dynamic full node config builder
Adding a full node network is similar to instantiating a validator config. The
difference is that if a config already exists at that path, the assumption is
to append the full node network, if the network isn't already defined.
Secondly, full nodes support public or permissioned networks. The former
requires no further configuration, whereas the latter needs a full node
specific seed, total number of full nodes, and index into the configuration
set. The total number includes the upstream peer which is indexed into the
first position (0).

    full-node-config-builder \
        -a $PUBLIC_MULTIADDR_FOR_NODE_I \
        -b $PUBLIC_MULTIADDR_FOR_NODE_0 \
        -d /opt/libra/data \
        -l $ANY_MULTIADDR_FOR_NODE_I \
        -n $TOTAL_NUMBER_OF_NODES \
        -o /opt/libra/etc \
        -s $SHARED_SECRET \
        [ -i $I -f $TOTAL_NUMBER_OF_FULL_NODES -c $FULL_NODE_SHARED_SECRET | -p ]

As an example a of adding 4 membered permissioned network connecting to the
node above:

    full-node-config-builder \
        -a "ip4/1.1.1.2/tcp/7100" \
        -b "ip4/1.1.1.2/tcp/7100" \
        -d /opt/libra/fn/data \
        -l "/ip4/0.0.0.0/tcp/7100" \
        -n 4
        -o /opt/libra/fn/etc \
        -s 0123456789abcdef101112131415161718191a1b1c1d1e1f2021222324252627 \
        -i 0 \
        -f 4 \
        -c 28292a2b2c2d2e2f303132333435363738393a3b3c3d3e3f4041424344454547

Similarly a public network could be added via:

    full-node-config-builder \
        -a "ip4/1.1.1.2/tcp/7100" \
        -b "ip4/1.1.1.2/tcp/7100" \
        -d /opt/libra/fn/data \
        -l "/ip4/0.0.0.0/tcp/7100" \
        -n 4
        -o /opt/libra/fn/etc \
        -s 0123456789abcdef101112131415161718191a1b1c1d1e1f2021222324252627 \
        -p

Closes: #1963
Approved by: davidiw