Integrating specification language support into ./language/move-prover.
This adapts concepts of David's spec language prototype and combines it with Sam's new spec parsing support in the Move parser. It also refactors and cleans up some of the tool level plumbing for the move prover, so we can more productively iterate on it.

The spec language support is not complete, as indicated the by TODO markers in the source.

Quite a bit work in this PR is about getting a better command line interface and test support for MVP (the Move Prover). I apologize to have not split this out. If you are mostly interested in the actual spec translation (and its open problems), see `src/spec_translator.rs`, and the related test output at `tests\goldenfiles\test-specs.bpl`.

Closes: #2044
Approved by: sblackshear