[temppath] Support custom temp directories
In order to perform certain operations, a temp path may need to be on a
specific file system. The only way to do this is to control where the
path is spawned. For example, the rename operation only works if the
destination is on the same file system as the source.

Closes: #2358
Approved by: aching