[logger] Move security log out of logger
There doesn't seem to be a clear reason to have these in the same crate.
In fact, one could argue that anything under common should be oblviious
to libra core concepts and security-logger should possibly live in
another path. Also this brings in backtrace, which we can now remove
from the dependencies for libra logger.

Closes: #2556
Approved by: mimoo