[language] split Move VM out into its own crates
One of the design goals for the Move VM is for it to be possible to use it
in other blockchains. This split is an important step towards achieving that
goal.

Introduce two crates:

* `move-vm-runtime`: core interpreter for the Move VM
* `move-vm-state`: an implementation of state management for the Move VM

There's still a lot more work to be done -- such as figuring out how
flexible things like account addresses can be.

Closes: #2694
Approved by: dariorussi