[language] introduce move-core-types and move Identifier into it
This crate will eventually contain all the types that are central to Move,
with the eventual goal being to release this on crates.io as a library for
others to consume.

Note that this library sits *underneath* libra-types (since the types here
are going to be independent of Libra).

Also start off the migration by moving the Identifier type over to this
crate.

Closes: #2827
Approved by: vgao1996