[move-prover] Setting up infrastructure for analysis/transformation tests.
This sets up the infrastrucure for testing analysis phases (which create annotations) and transformations.

- A printer for stackless bytecode functions.
- The printer supports a mechanism to dump annotations as produced by analysis phases.
- A test driver which executes baseline tests in `stackless-bytecode-generator/tests/<subdir>` where each `subdir` represents a set of tests related to a particular pipeline. These tests are based on printing out the initial bytecode and the bytecode+annotations for the result of running a pipeline.

This new feature is demonstrated on a basic test in `stackless-bytecode-generator/tests/lifetime`.

This PR also adds a new analysis for reaching definitions, and tests it using the above mechanism, but nothing is yet done with the result of this analysis. It also adds some code to `stackless_control_graph.rs` for tracking predecessors blocks in addition to successor blocks, which will allow backward analysis (in a future PR).

Closes: #3401
Approved by: shazqadeer